tlsn_core::config::tls_commit::mpc

Struct MpcTlsConfigBuilder

pub struct MpcTlsConfigBuilder { /* private fields */ }

Builder for MpcTlsConfig.

Implementations

impl MpcTlsConfigBuilder

pub fn max_sent_data(self, max_sent_data: usize) -> Self

Sets the maximum number of bytes that can be sent.

pub fn max_sent_records(self, max_sent_records: usize) -> Self

Sets the maximum number of application data records that can be sent.

pub fn max_recv_data_online(self, max_recv_data_online: usize) -> Self

Sets the maximum number of bytes that can be decrypted online.

pub fn max_recv_data(self, max_recv_data: usize) -> Self

Sets the maximum number of bytes that can be received.

pub fn max_recv_records_online(self, max_recv_records_online: usize) -> Self

Sets the maximum number of received application data records that can be decrypted online.

pub fn defer_decryption_from_start( self, defer_decryption_from_start: bool, ) -> Self

Sets whether the deferred decryption feature is toggled on from the start of the MPC-TLS connection.

pub fn network(self, network: NetworkSetting) -> Self

Sets the network settings.

pub fn build(self) -> Result<MpcTlsConfig, MpcTlsConfigError>

Builds the configuration.