Module tlsn_core::fixtures

Fixtures for testing

Modules

cert

Certificate fixtures

Functions

client_random

Returns a client random fixture.

encoder

Returns an encoder fixture.

encoder_seed

Returns an encoder seed fixture.

encoding_provider

Returns an encoding provider fixture using the given transcripts.

handshake_commitment

Returns a handshake commitment fixture.

handshake_data

Returns a handshake data fixture.

handshake_summary

Returns a handshake summary fixture.

notary_signing_key

Returns a notary signing key fixture.

server_cert_details

Returns a server certificate details fixture.

server_ephemeral_key

Returns a server ephemeral key fixture.

server_kx_details

Returns a server key exchange details fixture.

server_random

Returns a server random fixture.

session_header

Returns a session header fixture using the given transcript lengths and merkle root.