Module tlsn_core::proof

Different types of proofs used in the TLSNotary protocol.

Structs

SessionInfo

Contains information about the session

SessionProof

A session proof which is created from a crate::session::NotarizedSession

SubstringsProof

A substring proof using commitments

SubstringsProofBuilder

A builder for SubstringsProof

TlsProof

Proof that a transcript of communications took place between a Prover and Server.

Enums

SessionProofError

An error that can occur while verifying a SessionProof.

SubstringsProofBuilderError

An error for SubstringsProofBuilder

SubstringsProofError

An error relating to SubstringsProof

Functions

default_cert_verifier

Create a new [WebPkiVerifier] with the default trust anchors from the webpki-roots crate.