tlsn_core::transcript

Struct Idx

pub struct Idx(/* private fields */);

Transcript index.

Implementations

impl Idx

pub fn builder() -> IdxBuilder

Creates a new index builder.

pub fn empty() -> Self

Creates an empty index.

pub fn new(ranges: impl Into<RangeSet<usize>>) -> Self

Creates a new transcript index.

pub fn start(&self) -> usize

Returns the start of the index.

pub fn end(&self) -> usize

Returns the end of the index, non-inclusive.

pub fn iter(&self) -> impl Iterator<Item = usize> + '_

Returns an iterator over the values in the index.

pub fn iter_ranges(&self) -> impl Iterator<Item = Range<usize>> + '_

Returns an iterator over the ranges of the index.

pub fn len(&self) -> usize

Returns the number of values in the index.

pub fn is_empty(&self) -> bool

Returns whether the index is empty.

pub fn count(&self) -> usize

Returns the number of disjoint ranges in the index.