tlsn_core::transcript

Struct PartialTranscript

pub struct PartialTranscript { /* private fields */ }

A partial transcript.

A partial transcript is a transcript which may not have all the data authenticated.

Implementations

impl PartialTranscript

pub fn new(sent_len: usize, received_len: usize) -> Self

Creates a new partial transcript initalized to all 0s.

Arguments

sent_len - The length of the sent data.

received_len - The length of the received data.

pub fn len_sent(&self) -> usize

Returns the length of the sent transcript.

pub fn len_received(&self) -> usize

Returns the length of the received transcript.

pub fn is_complete(&self) -> bool

Returns whether the transcript is complete.

pub fn contains(&self, direction: Direction, idx: &RangeSet<usize>) -> bool

Returns whether the index is in bounds of the transcript.

pub fn sent_unsafe(&self) -> &[u8]

Returns a reference to the sent data.

Warning

Not all of the data in the transcript may have been authenticated. See sent_authed for a set of ranges which have been.

pub fn received_unsafe(&self) -> &[u8]

Returns a reference to the received data.

Warning

Not all of the data in the transcript may have been authenticated. See received_authed for a set of ranges which have been.

pub fn sent_authed(&self) -> &RangeSet<usize>

Returns the index of sent data which have been authenticated.

pub fn received_authed(&self) -> &RangeSet<usize>

Returns the index of received data which have been authenticated.

pub fn sent_unauthed(&self) -> RangeSet<usize>

Returns the index of sent data which havenโt been authenticated.

pub fn received_unauthed(&self) -> RangeSet<usize>

Returns the index of received data which havenโt been authenticated.

pub fn iter(&self, direction: Direction) -> impl Iterator<Item = u8> + '_

Returns an iterator over the authenticated data in the transcript.

pub fn union_transcript(&mut self, other: &PartialTranscript)

Unions the authenticated data of this transcript with another.

Panics

Panics if the other transcript is not the same length.

pub fn union_subsequence(&mut self, direction: Direction, seq: &Subsequence)

Unions an authenticated subsequence into this transcript.

Panics

Panics if the subsequence is outside the bounds of the transcript.

pub fn set_unauthed(&mut self, value: u8)

Sets all bytes in the transcript which havenโt been authenticated.

Arguments

value - The value to set the unauthenticated bytes to

pub fn set_unauthed_range( &mut self, value: u8, direction: Direction, range: Range<usize>, )

Sets all bytes in the transcript which havenโt been authenticated within the given range.

Arguments

value - The value to set the unauthenticated bytes to

range - The range of bytes to set