tlsn_core::transcript

Struct Transcript

pub struct Transcript { /* private fields */ }

A transcript contains the plaintext of all application data communicated between the Prover and the Server.

Implementations

impl Transcript

pub fn new(sent: impl Into<Vec<u8>>, received: impl Into<Vec<u8>>) -> Self

Creates a new transcript.

pub fn sent(&self) -> &[u8]

Returns a reference to the sent data.

pub fn received(&self) -> &[u8]

Returns a reference to the received data.

pub fn len(&self) -> (usize, usize)

Returns the length of the sent and received data, respectively.

pub fn length(&self) -> TranscriptLength

Returns the transcript length.

pub fn get( &self, direction: Direction, idx: &RangeSet<usize>, ) -> Option<Subsequence>

Returns the subsequence of the transcript with the provided index, returning None if the index is out of bounds.

pub fn to_partial( &self, sent_idx: RangeSet<usize>, recv_idx: RangeSet<usize>, ) -> PartialTranscript

Returns a partial transcript containing the provided indices.

Panics

Panics if the indices are out of bounds.

Arguments

sent_idx - The indices of the sent data to include.

recv_idx - The indices of the received data to include.