Module tlsn_core::request

Attestation requests.

After the TLS connection, a Prover can request an attestation from the Notary which contains various information about the connection. During this process the Prover has the opportunity to configure certain aspects of the attestation, such as which signature algorithm the Notary should use to sign the attestation. Or which hash algorithm the Notary should use to merkelize the fields.

A Request can be created using a RequestBuilder. The builder will take both configuration via a RequestConfig as well as the Prover’s secret data. The Secrets are of course not shared with the Notary but are used to create commitments which are included in the attestation.

Structs

InconsistentAttestation

Error for Request::validate.

Request

Attestation request.

RequestBuilder

Builder for Request.

RequestBuilderError

Error for RequestBuilder.

RequestConfig

Request configuration.

RequestConfigBuilder

Builder for RequestConfig.

RequestConfigBuilderError

Error for RequestConfigBuilder.