tlsn_core

Module transcript

Transcript types.

All application data communicated over a TLS connection is referred to as a Transcript. A transcript is essentially just two vectors of bytes, each corresponding to a Direction.

TLS operates over a bidirectional byte stream, and thus there are no application layer semantics present in the transcript. For example, HTTPS is an application layer protocol that runs over TLS so there is no concept of “requests” or “responses” in the transcript itself. These semantics must be recovered by parsing the application data and relating it to the bytes in the transcript.

Selective Disclosure

Using a TranscriptProof a Prover can selectively disclose parts of a transcript to a Verifier in the form of a PartialTranscript. A Verifier always learns the length of the transcript, but sensitive data can be withheld.

Modules

hash

Plaintext hash commitments.

Structs

CompressedPartialTranscript

PartialTranscript in a compressed form.

InvalidSubsequence

Invalid subsequence error.

PartialTranscript

A partial transcript.

Record

A TLS record.

Subsequence

Transcript subsequence.

TlsTranscript

A transcript of TLS records sent and received by the prover.

Transcript

A transcript contains the plaintext of all application data communicated between the Prover and the Server.

TranscriptCommitConfig

Configuration for transcript commitments.

TranscriptCommitConfigBuilder

A builder for TranscriptCommitConfig.

TranscriptCommitConfigBuilderError

Error for TranscriptCommitConfigBuilder.

TranscriptCommitRequest

Request to compute transcript commitments.

TranscriptProof

Proof of the contents of a transcript.

TranscriptProofBuilder

Builder for TranscriptProof.

TranscriptProofBuilderError

Error for TranscriptProofBuilder.

TranscriptProofError

Error for TranscriptProof.

Enums

ContentType

TLS record content type.

Direction

The direction of data communicated over a TLS connection.

TranscriptCommitment

Transcript commitment.

TranscriptCommitmentKind

Kind of transcript commitment.

TranscriptSecret

Secret for a transcript commitment.